+++ lib/duolingo4d.dart
@@ -8,6 +8,7 @@ export 'src/duolingo_request_header.dart';
 export 'src/duolingo_cookie_jar.dart';
 export 'src/response/response.dart';
 export 'src/response/status.dart';
+export 'src/response/manifest/manifest_response.dart';
 export 'src/response/versioninfo/version_info_response.dart';
 export 'src/response/auth/auth_response.dart';
 export 'src/response/user/user_response.dart';